+++ packages/Maleo.js/CHANGELOG.md
@@ -3,9 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.2-canary.0](https://github.com/airyrooms/maleo.js/compare/v0.1.1...v0.1.2-canary.0) (2019-04-04)
+# [0.2.0](https://github.com/airyrooms/maleo.js/compare/v0.1.1...v0.2.0) (2019-04-06)
 
-**Note:** Version bump only for package @airy/maleo
+
+### Bug Fixes
+
+* **maleo-core/server:** route handler promises ([#180](https://github.com/airyrooms/maleo.js/issues/180)) ([6720332](https://github.com/airyrooms/maleo.js/commit/6720332))
+
+
+### Features
+
+* **maleo-core/render:** changed the way we render wrap ([#181](https://github.com/airyrooms/maleo.js/issues/181)) ([0e160da](https://github.com/airyrooms/maleo.js/commit/0e160da))